server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
language-tool-python==2.7.1
requests==2.31.0
PyPDF2==3.0.1
Pillow==10.0.1
pdf2image==1.16.3
groq>=0.8.0
